backend/python/pyproject.toml
[tool.poetry]
name = "vracademyauditionserver"
version = "0.1.0"
description = "VR Academy Drawing Server"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.9.1"
aiomysql = "^0.2.0"
protobuf = ">=4.21.6,<5.0.0dev"
grpcio = ">=1.54.0,<1.55.0"
grpcio-tools = ">=1.54.0,<1.55.0"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
google-generativeai = "^0.5.4"
anthropic = "^0.20.0"
mistralai = "^0.1.3"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"